// src/app/models/machine.interface.ts
export interface MachineInterface {
    Usuario: string;
    Email: string;
    Nome_da_Maquina: string;
    Sistema_Operacional: string;
    Placa_mae: string;
    Processador: string;
    Armazenamento: string;
    Placa_de_Video: string;
    Conector_de_Rede: string;
    Quantidade_de_RAM: string;
    Geracao_da_RAM: string;
    Termo: string;
}